requirements.txt
alembic==1.12.0
anyio==4.0.0
Babel==2.12.1
blinker==1.6.2
Bottleneck==1.3.7
Brotli==1.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.4
dnspython==2.4.2
dominate==2.8.0
email-validator==2.0.0.post2
exceptiongroup==1.1.3
Flask==2.2.2
flask-babel==3.1.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.0.1
greenlet==2.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
langdetect==1.0.9
Mako==1.2.4
MarkupSafe==2.1.3
numexpr==2.8.4
numpy==1.26.0
packaging==23.1
pandas==2.0.3
pip==23.2.1
pycparser==2.21
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2023.3.post1
requests==2.31.0
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
visitor==0.1.3
Werkzeug==2.3.7
wheel==0.41.2
WTForms==3.0.1
zipp==3.17.0
